Rename county audit submodule.
